style: move the IsCompatible method to a new util.go file
